requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
fastapi==0.115.4
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
nodeenv==1.9.1
openai==1.52.2
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyright==1.1.386
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.41.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
youtube-transcript-api==0.6.2
